requirements.txt
Django==2.1
django-cors-headers==3.0.2
django-filter==2.1.0
django-heroku==0.3.1
dj-database-url==0.5.0
djangorestframework==3.9.4
psycopg2-binary==2.8.3
pytz==2019.1
whitenoise==4.1.2
gunicorn==19.9.0
